Renamed query manager.
