+++ src/persistence/primary/query/SingleResultQueryManager.ts
@@ -1,10 +1,10 @@
 import { Entity } from '../../../model/entity';
 import { VOID_RESULT_STRING } from '../LuaConstants';
 import { PersistencySearchOptions } from '../options/persistency-search-options';
+import { AntPrimaryQueryManager } from './ant-primary-query-manager';
 import { ISingleResultQueryManager } from './ISingleResultQueryManager';
-import { PrimaryQueryManager } from './primary-query-manager';
 
-export class SingleResultQueryManager<TEntity extends Entity> extends PrimaryQueryManager<TEntity, number | string>
+export class SingleResultQueryManager<TEntity extends Entity> extends AntPrimaryQueryManager<TEntity, number | string>
   implements ISingleResultQueryManager<TEntity> {
   /**
    * True if the queries managed can return multiple results.
